refact: accept contribution from Events and Projects
