+++ components/contribution-flow/StepProfileLoggedInForm.js
@@ -1,10 +1,6 @@
 import React, { Fragment, useEffect } from 'react';
 import PropTypes from 'prop-types';
-import { uniqBy } from 'lodash';
 import { FormattedMessage } from 'react-intl';
-
-import { CollectiveType } from '../../lib/constants/collectives';
-import roles from '../../lib/constants/roles';
 
 import { Box, Flex } from '../Grid';
 import PrivateInfoIcon from '../icons/PrivateInfoIcon';
@@ -17,47 +13,9 @@ import { useUser } from '../UserProvider';
 
 import ContributeProfilePicker from './ContributeProfilePicker';
 import StepProfileInfoMessage from './StepProfileInfoMessage';
-import { contributionRequiresAddress, contributionRequiresLegalName } from './utils';
+import { contributionRequiresAddress, contributionRequiresLegalName, getContributorProfiles } from './utils';
 
 export const NEW_ORGANIZATION_KEY = 'newOrg';
-
-const memberCanBeUsedToContribute = (member, account, canUseIncognito) => {
-  if (member.role !== roles.ADMIN) {
-    return false;
-  } else if (member.collective.id === account.legacyId) {
-    // Collective can't contribute to itself
-    return false;
-  } else if (!canUseIncognito && member.collective.isIncognito) {
-    // Incognito can't be used to contribute if not allowed
-    return false;
-  } else if (
-    member.collective.type === CollectiveType.COLLECTIVE &&
-    member.collective.host?.id !== account.host.legacyId
-  ) {
-    // If the contributing account is fiscally hosted, the host mush be the same as the one you're contributing to
-    return false;
-  } else if ([CollectiveType.EVENT, CollectiveType.PROJECT].includes(member.collective.type)) {
-    // Not supported yet, see https://github.com/opencollective/opencollective/issues/5066
-    // To enable contributions from them, we need to remove this check and make sure they're fetched from the API
-    // (they're currently not because we only fetch direct members, not parent's)
-    return false;
-  } else {
-    return true;
-  }
-};
-
-const getProfiles = (loggedInUser, collective, canUseIncognito) => {
-  if (!loggedInUser) {
-    return [];
-  } else {
-    const filteredMembers = loggedInUser.memberOf.filter(member =>
-      memberCanBeUsedToContribute(member, collective, canUseIncognito),
-    );
-    const memberProfiles = filteredMembers.map(member => member.collective);
-    const personalProfile = { email: loggedInUser.email, image: loggedInUser.image, ...loggedInUser.collective };
-    return uniqBy([personalProfile, ...memberProfiles], 'id');
-  }
-};
 
 const getDefaultProfile = (stepProfile, profiles, defaultProfileSlug) => {
   // If there's a default profile slug, enforce it
@@ -98,7 +56,7 @@ const getProfileInfo = (stepProfile, profiles) => {
 const StepProfileLoggedInForm = ({ defaultProfileSlug, onChange, canUseIncognito, collective, data, stepDetails }) => {
   const { LoggedInUser } = useUser();
   const getProfileArgs = [LoggedInUser, collective, canUseIncognito];
-  const profiles = React.useMemo(() => getProfiles(...getProfileArgs), getProfileArgs);
+  const profiles = React.useMemo(() => getContributorProfiles(...getProfileArgs), getProfileArgs);
   const defaultProfile = getDefaultProfile(data, profiles, defaultProfileSlug);
   const profileInfo = getProfileInfo(data, profiles);
