+++ lib/graphql/queries.js
@@ -162,6 +162,22 @@ export const loggedInUserQuery = gql`
             country
             structured
           }
+          children {
+            id
+            slug
+            type
+            name
+            isActive
+            imageUrl
+            host {
+              id
+            }
+            paymentMethods {
+              id
+              service
+              type
+            }
+          }
         }
       }
     }
